--- a/.github/workflows/vanilla-test.yml
+++ b/.github/workflows/vanilla-test.yml
@@ -5,9 +5,11 @@ on:
     branches:
       - master
     paths:
+      - Project.toml
       - test/environments/main/Project.toml
   pull_request:
     paths:
+      - Project.toml
       - test/environments/main/Project.toml
 
 jobs: