--- a/Project.toml
+++ b/Project.toml
@@ -17,7 +17,7 @@ Tables = "bd369af6-aec1-5ad0-b16a-f7cc5008161c"
 
 [compat]
 ArgCheck = "1, 2.0"
-BangBang = "0.3.18"
+BangBang = "0.3.20"
 InitialValues = "0.2.1"
 Requires = "0.5, 1.0"
 Setfield = "0.3, 0.4, 0.5, 0.6"