--- a/docs/Manifest.toml
+++ b/docs/Manifest.toml
@@ -7,7 +7,7 @@ version = "2.0.0"
 
 [[BangBang]]
 deps = ["Compat", "ConstructionBase", "Future", "InitialValues", "LinearAlgebra", "Requires", "Setfield", "Tables", "ZygoteRules"]
-git-tree-sha1 = "2b17ebdb66b1eac0c9c60ab1fa30476a309583e8"
+git-tree-sha1 = "1e63079e545ee3a988ba393025276766515c86a3"
 repo-rev = "master"
 repo-url = "https://github.com/tkf/BangBang.jl.git"
 uuid = "198e06fe-97b7-11e9-32a5-e1d131e6ad66"