* [ ] https://github.com/tkf/BangBang.jl/issues/141
* [x] Update compat