@tkf 2020-05-14
It's somewhat below the noise on GitHub Actions https://github.com/tkf/Transducers.jl/pull/274#issuecomment-628388167.

In Travis:

```
              ["groupby", "sum", "sac"]              1.10 (5%) :x:    1.00 (1%)
     ["groupby", "sum", "xf-with-init"] 0.74 (5%) :whitecheckmark:    1.00 (1%)
  ["groupby", "sum", "xf-without-init"] 0.74 (5%) :whitecheckmark:    1.00 (1%)
```

https://travis-ci.com/github/tkf/Transducers.jl/jobs/333583419#L876-L878

On my laptop:

Before:

```
1-element BenchmarkTools.BenchmarkGroup:
  tags: []
  "sum" => 3-element BenchmarkTools.BenchmarkGroup:
          tags: []
          "xf-without-init" => Trial(309.752 μs)
          "sac" => Trial(150.147 μs)
          "xf-with-init" => Trial(296.652 μs)
```

After:

```
1-element BenchmarkTools.BenchmarkGroup:
  tags: []
  "sum" => 3-element BenchmarkTools.BenchmarkGroup:
          tags: []
          "xf-without-init" => Trial(201.071 μs)
          "sac" => Trial(149.924 μs)
          "xf-with-init" => Trial(193.529 μs)
```